[u] Use sandbox as model for personal deployments (#2445)
