+++ deployments/sandbox/environment.py
@@ -3,6 +3,8 @@ from typing import (
     Optional,
 )
 
+is_sandbox = '/sandbox/' in __file__
+
 
 def env() -> Mapping[str, Optional[str]]:
     """
@@ -23,16 +25,23 @@ def env() -> Mapping[str, Optional[str]]:
     provide the value.
     """
     return {
-        # Set variables for the `sandbox` deployment here. The sandbox can be used to
-        # run integration tests against a PR and to perform CI/CD experiments.
+        # Set variables for the `sandbox` deployment here. The sandbox is used
+        # to run integration tests against PRs and to perform CI/CD experiments.
         #
-        # Only modify this file if you intend to commit those changes. To change the
-        # environment with a setting that's specific to you AND the deployment, create a
-        # environment.local.py right next to this file and make your changes there.
-        # Settings applicable to all environments but specific to you go into
-        # environment.local.py at the project root.
+        # You can use this file as a template for a personal deployment. Look
+        # for conditionals using the `is_sandbox` variable and adjust the `else`
+        # branch accordingly.
+        #
+        # Only modify this file if you intend to commit those changes. To apply
+        # a setting that's specific to you AND the deployment, create an
+        # `environment.local.py` file right next to this one and apply that
+        # setting there. Settings that are applicable to all environments but
+        # specific to you go into `environment.local.py` at the project root.
 
-        'AZUL_DEPLOYMENT_STAGE': 'sandbox',
+        # When using this file as a template for a personal deployment, replace
+        # `None` with a short string that is specific to to YOU.
+        #
+        'AZUL_DEPLOYMENT_STAGE': 'sandbox' if is_sandbox else None,
 
         'AZUL_CATALOGS': 'dcp2:repository/tdr:metadata/hca,'
                          'dcp2ebi:repository/tdr:metadata/hca,'
@@ -45,28 +54,43 @@ def env() -> Mapping[str, Optional[str]]:
         'AZUL_TDR_SERVICE_URL': 'https://jade.datarepo-dev.broadinstitute.org',
         'AZUL_SAM_SERVICE_URL': 'https://sam.dsde-dev.broadinstitute.org',
 
-        # The sandbox deployment uses a subdomain of the `dev` deployment's domain.
+        # This deployment uses a subdomain of the `dev` deployment's domain.
         #
         'AZUL_DOMAIN_NAME': 'dev.singlecell.gi.ucsc.edu',
         'AZUL_SUBDOMAIN_TEMPLATE': '*.{AZUL_DEPLOYMENT_STAGE}',
 
-        'AZUL_DRS_DOMAIN_NAME': 'drs.sandbox.dev.singlecell.gi.ucsc.edu',
+        'AZUL_DRS_DOMAIN_NAME': 'drs.{AZUL_DEPLOYMENT_STAGE}.dev.singlecell.gi.ucsc.edu',
 
         'AZUL_URL_REDIRECT_BASE_DOMAIN_NAME': 'dev.url.singlecell.gi.ucsc.edu',
         'AZUL_URL_REDIRECT_FULL_DOMAIN_NAME': '{AZUL_DEPLOYMENT_STAGE}.{AZUL_URL_REDIRECT_BASE_DOMAIN_NAME}',
 
-        # $0.186/h × 2 × 24h/d × 30d/mo = $267.84/mo
-        'AZUL_ES_INSTANCE_TYPE': 'r5.large.elasticsearch',
-        'AZUL_ES_INSTANCE_COUNT': '2',
+        **(
+            {
+                # $0.186/h × 2 × 24h/d × 30d/mo = $267.84/mo
+                'AZUL_ES_INSTANCE_TYPE': 'r5.large.elasticsearch',
+                'AZUL_ES_INSTANCE_COUNT': '2',
+            } if is_sandbox else {
+                # Personal deployments share an ES domain with `sandbox`
+                'AZUL_SHARE_ES_DOMAIN': '1',
+                'AZUL_ES_DOMAIN': 'azul-index-sandbox',
+                # Personal deployments use fewer Lambda invocations in parallel.
+                'AZUL_INDEXER_CONCURRENCY': '8',
+            }
+        ),
 
         'azul_dss_query_prefix': '42',
 
         'AZUL_DEBUG': '1',
 
-        'AZUL_OWNER': 'hannes@ucsc.edu',
+        # When using this file as a template for a personal deployment, change
+        # `None` to a string contaiing YOUR email address.
+        #
+        'AZUL_OWNER': 'hannes@ucsc.edu' if is_sandbox else None,
 
         'AZUL_AWS_ACCOUNT_ID': '122796619775',
         'AWS_DEFAULT_REGION': 'us-east-1',
 
+        # Set `GOOGLE_APPLICATION_CREDENTIALS` in `environment.local.py`
+        #
         'GOOGLE_PROJECT': 'platform-hca-dev',
     }
